Moving DAU logic to separate file outside scene delegate
