+++ App/iOS/Delegates/SceneDelegate.swift
@@ -44,6 +44,7 @@ class SceneDelegate: UIResponder, UIWindowSceneDelegate {
       scene: windowScene,
       braveCore: AppState.shared.braveCore,
       profile: AppState.shared.profile,
+      dau: AppState.shared.dau,
       diskImageStore: AppState.shared.diskImageStore,
       migration: AppState.shared.migration,
       rewards: AppState.shared.rewards,
@@ -422,6 +423,7 @@ extension SceneDelegate {
   private func createBrowserWindow(scene: UIWindowScene,
                                    braveCore: BraveCoreMain,
                                    profile: Profile,
+                                   dau: DAU,
                                    diskImageStore: DiskImageStore?,
                                    migration: Migration?,
                                    rewards: Brave.BraveRewards,
@@ -480,6 +482,7 @@ extension SceneDelegate {
     let browserViewController = BrowserViewController(
       windowId: windowId,
       profile: profile,
+      dau: dau,
       diskImageStore: diskImageStore,
       braveCore: braveCore,
       rewards: rewards,
@@ -571,58 +574,6 @@ extension SceneDelegate {
 extension SceneDelegate: UIViewControllerRestoration {
   public static func viewController(withRestorationIdentifierPath identifierComponents: [String], coder: NSCoder) -> UIViewController? {
     return nil
-  }
-}
-
-extension BrowserViewController {
-  func handleReferralLookup(_ urp: UserReferralProgram) {
-    if Preferences.URP.referralLookupOutstanding.value == true {
-      performProgramReferralLookup(urp, refCode: UserReferralProgram.getReferralCode())
-    } else {
-      urp.pingIfEnoughTimePassed()
-    }
-  }
-  
-  func handleSearchAdsInstallAttribution(_ urp: UserReferralProgram) {
-    urp.adCampaignLookup() { [weak self] response, error in
-      guard let self = self else { return }
-      
-      let refCode = self.generateReferralCode(attributionData: response, fetchError: error)
-      // Setting up referral code value
-      // This value should be set before first DAU ping
-      Preferences.URP.referralCode.value = refCode
-      Preferences.URP.installAttributionLookupOutstanding.value = false
-    }
-  }
-  
-  private func generateReferralCode(attributionData: AdAttributionData?, fetchError: Error?) -> String {
-    // Prefix code "001" with BRV for organic iOS installs
-    var referralCode = "BRV001"
-    
-    if fetchError == nil, attributionData?.attribution == true, let campaignId = attributionData?.campaignId {
-      // Adding ASA User refcode prefix to indicate
-      // Apple Ads Attribution is true
-      referralCode = "ASA\(String(campaignId))"
-    }
-    
-    return referralCode
-  }
-  
-  private func performProgramReferralLookup(_ urp: UserReferralProgram, refCode: String?) {
-    urp.referralLookup(refCode: refCode) { referralCode, offerUrl in
-      // Attempting to send ping after first urp lookup.
-      // This way we can grab the referral code if it exists, see issue #2586.
-      if Preferences.URP.installAttributionLookupOutstanding.value == false {
-        AppState.shared.dau.sendPingToServer()
-      }
-      let retryTime = AppConstants.buildChannel.isPublic ? 1.days : 10.minutes
-      let retryDeadline = Date() + retryTime
-
-        Preferences.NewTabPage.superReferrerThemeRetryDeadline.value = retryDeadline
-
-      guard let url = offerUrl?.asURL else { return }
-      self.openReferralLink(url: url)
-    }
   }
 }
