+++ Sources/Brave/Frontend/Browser/BrowserViewController/BVC+BraveTalk.swift
