+++ Sources/Brave/Frontend/Browser/BrowserViewController/BVC+FindInPageDelegate.swift
