+++ Sources/Brave/Frontend/Browser/BrowserViewController/BVC+IPFSScheme.swift
