+++ Sources/Brave/Frontend/Browser/BrowserViewController/BVC+KeyCommands.swift
