+++ Sources/Brave/Frontend/Browser/BrowserViewController/BVC+OpenSearch.swift
