+++ Sources/Brave/Frontend/Browser/BrowserViewController/BVC+P3A.swift
