+++ Sources/Brave/Frontend/Browser/BrowserViewController/BVC+ProductNotification.swift
