+++ Sources/Brave/Frontend/Browser/BrowserViewController/BVC+ReaderMode.swift
