+++ Sources/Brave/Frontend/Browser/BrowserViewController/BVC+TabManagerDelegate.swift
