+++ Sources/Brave/Frontend/Browser/BrowserViewController/BVC+ToolbarDelegate.swift
