+++ Sources/Brave/Frontend/Browser/BrowserViewController/BVC+WKNavigationDelegate.swift
