+++ Sources/Brave/Frontend/Browser/BrowserViewController/BVC+Wallet.swift
