+++ Sources/Brave/Frontend/Browser/BrowserViewController/BVC+Web3NameService.swift
