+++ Sources/Brave/Frontend/Browser/BrowserViewController/BVC+Widgets.swift
