+++ Sources/Brave/Frontend/Browser/BrowserViewController/BrowserViewController.swift
@@ -162,6 +162,7 @@ public class BrowserViewController: UIViewController {
 
   public let windowId: UUID
   let profile: Profile
+  let dau: DAU
   let braveCore: BraveCoreMain
   let tabManager: TabManager
   let migration: Migration?
@@ -272,6 +273,7 @@ public class BrowserViewController: UIViewController {
   public init(
     windowId: UUID,
     profile: Profile,
+    dau: DAU,
     diskImageStore: DiskImageStore?,
     braveCore: BraveCoreMain,
     rewards: BraveRewards,
@@ -282,6 +284,7 @@ public class BrowserViewController: UIViewController {
   ) {
     self.windowId = windowId
     self.profile = profile
+    self.dau = dau
     self.braveCore = braveCore
     self.bookmarkManager = BookmarkManager(bookmarksAPI: braveCore.bookmarksAPI)
     self.rewards = rewards
