+++ Sources/Brave/Frontend/Browser/Search/OpenSearchEngineButton.swift
@@ -78,3 +78,4 @@ class OpenSearchEngineButton: BraveButton {
     }
   }
 }
+
